Factories folders created. Unnecessary tests deleted.
