+++ final_project/snapvisite/templates/snapvisite/schedule.html
@@ -6,7 +6,6 @@
     <div class="row">
         <div class="col-12 border-bottom border-end">
             <h3 class="display-4 text-center">Editor</h3>
-            <h5>Use delete if u want remove day.</h5>
         </div>
         <div class="col-3"></div>
         <div class="col-6 border-end border-start">
